% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get68proc.R
\name{get68proc}
\alias{get68proc}
\title{obtain a SingleCellExperiment with a version of TENxPBMCData pbmc68k data}
\usage{
get68proc(cache = BiocFileCache::BiocFileCache(), targetfolder = tempdir())
}
\arguments{
\item{cache}{a cache like that defined by BiocFileCache::BiocFileCache()}

\item{targetfolder}{character(1) path, defaulting to `tempdir()`, where a serialized
SingleCellExperiment will be placed}
}
\description{
obtain a SingleCellExperiment with a version of TENxPBMCData pbmc68k data
}
\note{
An 835MB zipfile holding the output of saveHDF5SummarizedExperiment is retrieved
from Open Storage Network and placed in cache.
colData labels are obtained with a SingleR run using the HumanPrimaryCellAtlasData
label.main.
}
\examples{
tst = get68proc()
dir.exists(file.path(tempdir(), "p68_saved"))
tst
table(tst$labels)
rhdf5::h5ls(slot(DelayedArray::seed(assay(tst)), "filepath"))
}
